requirements.txt
flask==3.1.0
anthropic==0.43.0
requests==2.32.3
python-dotenv==1.0.1
sqlalchemy==2.0.36
msal==1.31.1
